+++ src/Prettychart/Server.hs
@@ -14,15 +14,18 @@ import Box
 import Chart
 import Control.Concurrent.Async
 import Control.Monad (when)
-import Lucid as L
-import Optics.Core
+import Data.ByteString (ByteString)
+import Data.ByteString.Char8 (pack)
+import Data.Text.Encoding
+import MarkupParse
+import Optics.Core hiding (element)
 import Prettychart.Any
 import Web.Rep
 
 -- | 'Page' containing a web socket and javascript needed to run it.
-chartSocketPage :: Page
-chartSocketPage =
-  bootstrap5Page
+chartSocketPage :: Maybe ByteString -> Page
+chartSocketPage title =
+  bootstrapPage
     & #jsOnLoad
       .~ mconcat
         [ webSocket,
@@ -30,12 +33,7 @@ chartSocketPage =
           refreshJsbJs
         ]
     & #htmlBody
-      .~ divClass_
-        "container"
-        ( mconcat
-            [ divClass_ "row" $ divClass_ "col" (h4_ "prettychart" <> L.with div_ [id_ "prettychart"] mempty)
-            ]
-        )
+      .~ element "div" [Attr "class" "container"] (element "row" [Attr "class" "col"] (maybe mempty (elementc "h4" []) title) <> element_ "div" (pure $ Attr "id" "prettychart"))
 
 -- | Print a chart supplying a 'ChartOptions' consumer, and a showable thing that may be chartable. The first argument flags whether to also print the item to stdout.
 printChart :: (Show a) => Bool -> (ChartOptions -> IO Bool) -> a -> IO ()
@@ -50,14 +48,14 @@ printChart reprint send s = case anyChart (show s) of
 -- An iconic ghci session transcript:
 --
 -- >> import Chart.Examples
--- >> (sendChart, quitChartServer) <- startChartServer
+-- >> (sendChart, quitChartServer) <- startChartServer (Just "prettychart")
 -- >> sendChart unitExample
 --
 -- ... point browser to localhost:9160 ...
 --
 -- >> quitChartServer
-startChartServer :: IO (ChartOptions -> IO Bool, IO ())
-startChartServer = startChartServerWith defaultSocketConfig chartSocketPage
+startChartServer :: Maybe String -> IO (ChartOptions -> IO Bool, IO ())
+startChartServer title = startChartServerWith defaultSocketConfig (chartSocketPage $ pack <$> title)
 
 -- | Start the chart server protocol with bespoke 'SocketConfig' and 'Page' configurations.
 --
@@ -65,5 +63,5 @@ startChartServer = startChartServerWith defaultSocketConfig chartSocketPage
 startChartServerWith :: SocketConfig -> Page -> IO (ChartOptions -> IO Bool, IO ())
 startChartServerWith scfg page = do
   (Box c e, q) <- toBoxM Single
-  x <- async $ serveSocketBox scfg page (Box toStdout (replace "prettychart" . renderChartOptions <$> e))
+  x <- async $ serveSocketBox scfg page (Box toStdout (decodeUtf8 . replace "prettychart" . encodeChartOptions <$> e))
   pure (commit c, cancel x >> q)