+++ ChangeLog.md
@@ -1,3 +1,9 @@
+0.1.2
+===
+
+- removed example data and related charts
+- refactored to markup-parse
+
 0.1.1
 ===